Shader hot-reloading; Remove all old type declarations for scalar types

+++ kraft/src/core/kraft_memory.cpp
@@ -23,7 +23,7 @@ KRAFT_API MemoryBlock MallocBlock(uint64_t size, MemoryTag tag, bool Zero)
         MemZero(region, size);
     }
 
-    return { (uint8*)region, size, tag };
+    return { (u8*)region, size, tag };
 }
 
 KRAFT_API MemoryBlock ReallocBlock(MemoryBlock block, uint64_t size)
@@ -32,7 +32,7 @@ KRAFT_API MemoryBlock ReallocBlock(MemoryBlock block, uint64_t size)
     g_MemoryStats.AllocationsByTag[block.Tag] += size - block.Size;
 
     void* region = Platform::Realloc(block.Data, size, false);
-    block.Data = (uint8*)region;
+    block.Data = (u8*)region;
     block.Size = size;
 
     return block;
@@ -65,7 +65,7 @@ KRAFT_API void* Malloc(uint64_t size, MemoryTag tag, bool zero)
     return retval;
 }
 
-KRAFT_API void* Realloc(void* region, uint64 size, MemoryTag tag)
+KRAFT_API void* Realloc(void* region, u64 size, MemoryTag tag)
 {
     g_MemoryStats.Allocated += size;
     g_MemoryStats.AllocationsByTag[tag] += size;
@@ -73,7 +73,7 @@ KRAFT_API void* Realloc(void* region, uint64 size, MemoryTag tag)
     return Platform::Realloc(region, size, false);
 }
 
-KRAFT_API void* Realloc(void* region, uint64 size)
+KRAFT_API void* Realloc(void* region, u64 size)
 {
     return Platform::Realloc(region, size, false);
 }
@@ -91,7 +91,7 @@ KRAFT_API void Free(void* region, uint64_t size, MemoryTag tag)
     Platform::Free(region);
 }
 
-KRAFT_API void MemZero(void* region, uint64 size)
+KRAFT_API void MemZero(void* region, u64 size)
 {
     Platform::MemZero(region, size);
 }
@@ -101,7 +101,7 @@ KRAFT_API void MemZero(MemoryBlock block)
     Platform::MemZero(block.Data, block.Size);
 }
 
-KRAFT_API void MemSet(void* region, int value, uint64 size)
+KRAFT_API void MemSet(void* region, int value, u64 size)
 {
     Platform::MemSet(region, value, size);
 }
@@ -111,17 +111,17 @@ KRAFT_API void MemSet(MemoryBlock block, int value)
     MemSet(block.Data, value, block.Size);
 }
 
-KRAFT_API void MemCpy(void* dst, void const* src, uint64 size)
+KRAFT_API void MemCpy(void* dst, void const* src, u64 size)
 {
     Platform::MemCpy(dst, src, size);
 }
 
-KRAFT_API void MemCpy(MemoryBlock dst, MemoryBlock src, uint64 size)
+KRAFT_API void MemCpy(MemoryBlock dst, MemoryBlock src, u64 size)
 {
     MemCpy(dst.Data, src.Data, size);
 }
 
-KRAFT_API int MemCmp(const void* a, const void* b, uint64 size)
+KRAFT_API int MemCmp(const void* a, const void* b, u64 size)
 {
     return Platform::MemCmp(a, b, size);
 }
@@ -156,9 +156,9 @@ KRAFT_API MemoryStatsOutput GetMemoryStats()
 {
     MemoryStatsOutput Output;
 
-    const size_t kib = 1024;
-    const size_t mib = kib * kib;
-    const size_t gib = mib * kib;
+    const u64 kib = 1024;
+    const u64 mib = kib * kib;
+    const u64 gib = mib * kib;
 
     Output.AllocatedBytes = g_MemoryStats.Allocated;
     Output.AllocatedKiB = g_MemoryStats.Allocated / (float)kib;
